Add initial project code
OVERVIEW

Add current version of project code for a new `roots` package.

This project provides:

- the `roots` package for use by Go client code
  - (embedded) hashes and root certificates sourced from the Mozilla
    Root Program
- PEM certificate bundles for use with other applications (e.g., curl,
  wget, etc.)

The `roots` package provides:

- valid CA (root) certificates in the Mozilla Root Program
- removed CA (root) certificates
- upcoming CA (root) certificate removals
- certificate hashes for all sets

The audience for the functionality provided by the `roots` package is
primarily diagnostic tools which need direct access to
`x509.Certificate` values for certificate chain analysis purposes and
not general use applications/tooling.

USE CASES

The provided root certificates and root certificate hashes are
intended for use by diagnostic tools when evaluating certificate
chains for common misconfiguration issues.

The collections of root certificate hashes are useful for identifying
which of the the provided collections an existing root certificate
belongs to.

NOTE

The `count.go` file is generated when the `generate.go` code is
executed to regenerate project assets. This commit includes a copy of
this file populated with current counts of the entries for each
provided collection of certificates and hashes. This commit does *NOT*
include a copy of those generated assets however.

The goal is to have those assets added via an automated commit from
GitHub Actions as a follow-up to this PR. The intent is to provide the
initial version of those resources in an easily verified way (though
admittedly the GH Actions logs are limited to 90 days).

CREDIT

The following LLM sources were used for inspiration or starting code
samples for some of the included changes:

- ChatGPT, OpenAI
- Google Gemini
- Claude (Anthropic AI assistant)

All were helpful in pointing me in the right direction when I needed
it, particularly function naming patterns.

NOTE: This project is based heavily on the `atc0005/intermediates`
project and the `FiloSottile/intermediates` project that is forked
from. Much of the LLM assistance noted here was provided when setting
up the `atc0005/intermediates` project for active use (and replicated
here).

REFERENCES

- https://github.com/atc0005/intermediates
- https://www.ccadb.org/rootstores/usage#ccad
- https://wiki.mozilla.org/CA/Included_Certificates
- https://wiki.mozilla.org/CA/Removed_Certificates