Update docs to note Go 1.23 requirements
OVERVIEW

Go 1.23 changed the default behavior of `crypto/x509.ParseCertificate`
to reject serial numbers that are negative. Because some of the root
certificates in this collection contain negative serial numbers the
consuming application will need to explicitly enable the
`x509negativeserial` GODEBUG setting *OR* opt to use the certificate
hashes for removed root certificates instead of affected collections.

CHANGES

- update README file to note Go 1.23 requirements
- update doc comments for affected functions to make this requirement
  clear
- fix instances of "intermediates" in function doc comments to say
  "certificates" instead (missed copy/paste/modify work when
  bootstrapping this project)

REFERENCES

- https://pkg.go.dev/crypto/x509#ParseCertificate
- https://golang.org/doc/godebug
- https://stackoverflow.com/questions/79061981